+++ ndsampler/coco_frames.py
@@ -23,7 +23,7 @@ class CocoFrames(abstract_frames.Frames, util_misc.HashIdentifiable):
         >>> import ndsampler
         >>> import kwcoco
         >>> import ubelt as ub
-        >>> workdir = ub.ensure_app_cache_dir('ndsampler')
+        >>> workdir = ub.Path.appdir('ndsampler').ensuredir()
         >>> dset = kwcoco.CocoDataset.demo(workdir=workdir)
         >>> dset._ensure_imgsize()
         >>> self = CocoFrames(dset, workdir=workdir)
@@ -81,11 +81,11 @@ class CocoFrames(abstract_frames.Frames, util_misc.HashIdentifiable):
             >>> sampler2 = ndsampler.CocoSampler.demo('vidshapes5-multispectral')
             >>> self = sampler1.frames
             >>> pathinfo = self._build_pathinfo(1)
-            >>> print('pathinfo = {}'.format(ub.repr2(pathinfo, nl=3)))
+            >>> print('pathinfo = {}'.format(ub.urepr(pathinfo, nl=3)))
 
             >>> self = sampler2.frames
             >>> pathinfo = self._build_pathinfo(1)
-            >>> print('pathinfo = {}'.format(ub.repr2(pathinfo, nl=3)))
+            >>> print('pathinfo = {}'.format(ub.urepr(pathinfo, nl=3)))
         """
         img = self.dset.imgs[image_id]
         default_fname = img.get('file_name', None)